requirements.txt
Django~=2.2
djangorestframework~=3.11.0
